+++ nominal/src/lib.rs
@@ -6,3 +6,4 @@ pub use config::{Config, Profile};
 pub use core::{NominalClient, NominalClientBuilder};
 pub use error::{Error, Result};
 pub use rustls::client::ResolvesClientCert;
+pub use nominal_streaming as streaming;