+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.3](https://github.com/nominal-io/nominal-client-rs/compare/nominal-v0.4.2...nominal-v0.4.3) - 2026-05-21
+
+### Added
+
+- add channel delimiter in dataset create ([#70](https://github.com/nominal-io/nominal-client-rs/pull/70))
+
+### Fixed
+
+- nominal_url method returns a valid url ([#72](https://github.com/nominal-io/nominal-client-rs/pull/72))
+
+## [0.4.2](https://github.com/nominal-io/nominal-client-rs/compare/nominal-v0.4.1...nominal-v0.4.2) - 2026-05-20
+
+### Added
+
+- add dataset with tag filter to an asset ([#67](https://github.com/nominal-io/nominal-client-rs/pull/67))
+
+## [0.4.1](https://github.com/nominal-io/nominal-client-rs/compare/nominal-v0.4.0...nominal-v0.4.1) - 2026-05-18
+
+### Added
+
+- add & re-export nominal-streaming ([#63](https://github.com/nominal-io/nominal-client-rs/pull/63))
+
+## [0.4.0](https://github.com/nominal-io/nominal-client-rs/compare/nominal-v0.3.0...nominal-v0.4.0) - 2026-05-18
+
+### Added
+
+- run create/search, asset create/search ([#38](https://github.com/nominal-io/nominal-client-rs/pull/38))
+
+### Fixed
+
+- use series metadata rather than channel metadata service for upsert channel info ([#60](https://github.com/nominal-io/nominal-client-rs/pull/60))
+
 ## [0.3.0](https://github.com/nominal-io/nominal-client-rs/compare/nominal-v0.2.3...nominal-v0.3.0) - 2026-04-29
 
 ### Added